+++ tesys/src/astrometry/src/hms.rs
@@ -4,7 +4,7 @@ use HOUR_PER_RAD;
 pub struct HMS {
     pub h: i8,
     pub m: i8,
-    pub s: f32,
+    pub s: f64,
     pub neg: bool,
 }
 
@@ -12,10 +12,10 @@ impl HMS {
     /*!
      * A container that can be used to express angles as hours, minutes and seconds.
      * The hours and minutes are represented as i8, whereas seconds has a fractional part
-     * thus is represented by an f32.
+     * thus is represented by an f64.
      */
 
-    pub fn new(_h: i8, _m: i8, _s: f32, _neg: bool) -> HMS {
+    pub fn new(_h: i8, _m: i8, _s: f64, _neg: bool) -> HMS {
         HMS {
             h: _h,
             m: _m,
@@ -24,10 +24,10 @@ impl HMS {
         }
     }
 
-    pub fn new_from_rad(&rad: &f32) -> HMS {
+    pub fn new_from_rad(&rad: &f64) -> HMS {
         /*!
-         * Creates a new instance of HMS object from a radian measurement, given as an f32.
-         * As the Angle struct uses an f32 representation of radians internally, this is what
+         * Creates a new instance of HMS object from a radian measurement, given as an f64.
+         * As the Angle struct uses an f64 representation of radians internally, this is what
          * gets called when you call the ```Angle::to_hms()``` function.
          */
 
@@ -41,7 +41,7 @@ impl HMS {
         };
         let _h: i8 = dummy.floor() as i8;
         let _m: i8 = (dummy.fract() * 60.0).floor() as i8;
-        let _s: f32 = (dummy.fract() * 60.0).fract() * 60.;
+        let _s: f64 = (dummy.fract() * 60.0).fract() * 60.;
 
         HMS {
             h: _h,
@@ -61,7 +61,7 @@ impl HMS {
 
         let _h: i8 = 0;
         let _m: i8 = 0;
-        let _s: f32 = 0.0;
+        let _s: f64 = 0.0;
         let _sign: bool = false;
         HMS::new(_h, _m, _s, _sign)
     }
@@ -74,8 +74,8 @@ impl fmt::Display for HMS {
     }
 }
 
-impl Into<f32> for HMS {
-    fn into(self) -> f32 {
-        (self.h as f32 + (self.m as f32) / 60.0 + (self.s as f32) / 3600.0) / HOUR_PER_RAD
+impl Into<f64> for HMS {
+    fn into(self) -> f64 {
+        (self.h as f64 + (self.m as f64) / 60.0 + (self.s as f64) / 3600.0) / HOUR_PER_RAD
     }
 }
